apps/api/requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.1
django-filter>=24.2

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Async & Channels
channels>=4.0.0
channels-redis>=4.2.0

# Celery
celery>=5.3.4
django-celery-beat>=2.5.0
redis>=5.0.0

# Storage (S3)
django-storages>=1.14.2
boto3>=1.34.21

# Authentication & Security
django-allauth>=0.57.0

# API Documentation
drf-spectacular>=0.27.0

# PDF Generation
WeasyPrint>=60.2

# Utilities
python-decouple>=3.8
Pillow>=10.2.0
python-dateutil>=2.8.2
requests>=2.31.0

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==23.12.1
isort==5.13.2
ruff==0.1.9
pre-commit==3.6.0

# Monitoring
sentry-sdk[django]==1.39.2
